Create Ask Question component

+++ app/Http/Controllers/QuestionsController.php
@@ -81,7 +81,15 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     { 
+        $this->authorize("update", $question);
         $question->update($request->only('title', 'body'));
+
+        if($request->expectsJson()){
+            return response()->json([
+                'message' => 'Your question has been updated successfully',
+                'body_html' => $question->body_html
+            ]);
+        }
 
         return redirect('/questions')->with('success', 'Your question has been updated successfully');
     }
@@ -94,7 +102,14 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     { 
+        $this->authorize("delete", $question);
         $question->delete();
+
+        if($request->expectsJson()){
+            return response()->json([
+                'message' => 'Question is deleted successfully'
+            ]);
+        }
 
         return \redirect('/questions')->with('success', 'Question is deleted successfully');
     }
